refactor: ✨ : parachain relayer -> solochain relayer (#14)
This PR refactors the existing `parachain` relayer to make it a
`solochain` relayer, stripping it from everything related to Parachains,
Para IDs and the notion of a Relaychain (except naming-wise, since we
want to keep changes compared to upstream at a minimum). As such:
- The `parachain` relayer is now launched with the `solochain` flag
instead, for clarity.
- The parachain heads root field from the BEEFY leafs has been renamed
to BeefyExtraField, to make it more generic.
- The mechanism to generate parachain heads proof has been changed to
generate message commitment roots.
- Logs in general have been improved, with extra debug logs and improved
clarity of info/warn/error logs, to make it easier for relayer operators
to know what's happening under the hood (and easier for us to debug it).
- Comments have been added to a few key functions so it's easier to
understand what they do.